[SSHD-698] Remove unnecessary boxing / unboxing

+++ sshd-core/src/main/java/org/apache/sshd/client/session/ClientSession.java
@@ -217,7 +217,7 @@ public interface ClientSession
             }
 
             Integer exitStatus = channel.getExitStatus();
-            if ((exitStatus != null) && (exitStatus.intValue() != 0)) {
+            if ((exitStatus != null) && (exitStatus != 0)) {
                 throw new RemoteException("Remote command failed (" + exitStatus + "): " + command, new ServerException(exitStatus.toString()));
             }
