+++ sshd-core/src/main/java/org/apache/sshd/client/subsystem/sftp/SftpVersionSelector.java
@@ -174,7 +174,7 @@ public interface SftpVersionSelector {
                         }
 
                         for (Integer avail : available) {
-                            if (version == avail.intValue()) {
+                            if (version == avail) {
                                 return version;
                             }
                         }
